requirements.txt
fastapi
uvicorn[standard]
flake8
black
pre-commit
pydantic
pyyaml
pytest-cov
